Split joy pub from rosgui, organized gui prefabs
