+++ Runtime/Prefabs/ROS/ROSSingletons.prefab
@@ -11,8 +11,9 @@ GameObject:
   - component: {fileID: 2970021653359921310}
   - component: {fileID: 1662493632097492073}
   - component: {fileID: 261061624273665792}
+  - component: {fileID: 1049777862564182728}
   m_Layer: 0
-  m_Name: ROSClock
+  m_Name: ROSSingletons
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -61,3 +62,16 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   frequency: 1
+--- !u!114 &1049777862564182728
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6357952754565123669}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 72e340880f8dbff49a22db0cce1570b4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  JoyTopic: /joy
