+++ Runtime/Prefabs/SmarcGUI/Components.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 712247ab9e9ff854ab26b12b0d427bd1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
