+++ Runtime/Prefabs/SmarcGUI/Components/Label-TextField Narrow Variant.prefab.meta
