+++ Runtime/Scripts/VehicleComponents/ROS/Publishers/Joy_Pub.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 72e340880f8dbff49a22db0cce1570b4
